add errors and code refactor
