+++ rpc/src/error.rs
@@ -1,6 +1,7 @@
 use ckb_error::{AnyError, Error as CKBError, ErrorKind, InternalError, InternalErrorKind};
 use ckb_tx_pool::error::Reject;
 use jsonrpc_core::{Error, ErrorCode, Value};
+use schemars::JsonSchema;
 use std::fmt::{Debug, Display};
 
 /// CKB RPC error codes.
@@ -28,7 +29,7 @@ use std::fmt::{Debug, Display};
 ///
 /// Unless otherwise noted, all the errors return optional detailed information as `string` in the error
 /// object `data` field.
-#[derive(Debug, PartialEq, Clone, Copy, Eq)]
+#[derive(Debug, PartialEq, Clone, Copy, Eq, JsonSchema)]
 pub enum RPCError {
     /// (-1): CKB internal errors are considered to never happen or only happen when the system
     /// resources are exhausted.
