+++ rpc/src/gen_doc/main.rs
@@ -1,8 +1,7 @@
 //! this is a tool to generate rpc doc
-mod utils;
-use crate::utils::*;
+mod gen;
+use crate::gen::RpcDocGenerator;
 use ckb_rpc::module::*;
-use serde_json::Value;
 use std::fs;
 
 fn dump_openrpc_json() -> Result<(), Box<dyn std::error::Error>> {
@@ -23,7 +22,7 @@ fn dump_openrpc_json() -> Result<(), Box<dyn std::error::Error>> {
     dump("integration_test_rpc_doc.json", integration_test_rpc_doc())?;
     dump("indexer_rpc_doc.json", indexer_rpc_doc())?;
     dump("experiment_rpc_doc.json", experiment_rpc_doc())?;
-    eprintln!("dump openrpc json...");
+    eprintln!("finished dump openrpc json...");
     Ok(())
 }
 
@@ -41,75 +40,10 @@ fn gen_rpc_readme(readme_path: &str) -> Result<(), Box<dyn std::error::Error>> {
         indexer_rpc_doc(),
         experiment_rpc_doc(),
     ];
-    let mut rpc_module_methods = vec![];
-    for rpc in all_rpc {
-        if let serde_json::Value::Object(map) = rpc {
-            let module_title = map["info"]["title"].as_str().unwrap();
-            // strip `_rpc` suffix
-            let module_title = &module_title[..module_title.len() - 4];
-            let module_methods = map["methods"].as_array().unwrap();
-            let types = map["components"]["schemas"].as_object().unwrap();
-            rpc_module_methods.push(RpcModule {
-                module_title: module_title.to_owned(),
-                module_methods: module_methods.to_owned(),
-                types: types.to_owned(),
-            });
-        }
-    }
 
-    // sort rpc_module_methods accoring to module_title
-    rpc_module_methods.sort_by(|a, b| a.module_title.cmp(&b.module_title));
+    let generator = RpcDocGenerator::new(&all_rpc, readme_path.to_owned());
+    fs::write(readme_path, generator.gen_markdown())?;
 
-    let mut res = String::new();
-
-    // strip lines below `**NOTE:**`
-    let readme = fs::read_to_string(readme_path).unwrap_or("".to_owned());
-    let lines = readme.lines().collect::<Vec<_>>();
-    for &line in lines.iter() {
-        res.push_str(&(line.to_owned() + "\n"));
-        if line.contains("**NOTE:** the content below is generated by gen_doc") {
-            break;
-        }
-    }
-
-    let mut types: Vec<(String, &Value)> = vec![];
-    for rpc_module in rpc_module_methods.iter() {
-        for (name, ty) in rpc_module.types.iter() {
-            if !types.iter().any(|(n, _)| *n == *name) {
-                types.push((name.to_owned(), ty));
-            }
-        }
-    }
-
-    // generate methods menu
-    res.push_str("\n\n* [RPC Methods](#rpc-methods)\n");
-    for rpc_module in rpc_module_methods.iter() {
-        res.push_str(&rpc_module.gen_methods_menu());
-    }
-
-    // generate type menu
-    types.sort_by(|(name1, _), (name2, _)| name1.cmp(name2));
-    res.push_str("* [RPC Types](#rpc-types)\n");
-    for (name, _) in types.iter() {
-        let ty = format!(
-            "    * [Type `{}`](#type-{})\n",
-            capitlize(name),
-            name.to_lowercase()
-        );
-        res.push_str(&ty);
-    }
-
-    // generate methods content
-    for rpc_module in rpc_module_methods.iter() {
-        let content = format!("{}\n", rpc_module.gen_methods_content());
-        res.push_str(&content);
-    }
-
-    // generate type content
-    res.push_str("## RPC Types\n");
-    gen_type_content(&mut res, types);
-
-    fs::write(readme_path, res)?;
     Ok(())
 }
